Refactoring of jobs and services gems
Extraction of jobs and services Ruby code into appropriate
torquebox-jobs and torquebox-services gems.

Additionally [TORQUE-1075] Use service name in :inspect implementation
was fixed by implementing the inscpect method in the service, since
JRuby did not a good job at creating the object inspection on its own.
