+++ gems/services/lib/torquebox/services.rb
@@ -15,14 +15,8 @@
 # Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 # 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 
-module TorqueBox
+begin
+  require 'torquebox-services'
+rescue LoadError=>e
+  # ignore!
 end
-
-#require 'torquebox/kernel'
-require 'torquebox/service_registry'
-require 'torquebox/component_manager'
-require 'torquebox/injectors'
-require 'torquebox/logger'
-require 'torquebox/msc'
-
-
