+++ gems/torquebox/lib/torquebox.rb
@@ -18,6 +18,8 @@
 require 'torquebox-core'
 require 'torquebox-naming'
 require 'torquebox-messaging'
+require 'torquebox-services'
+require 'torquebox-jobs'
 require 'torquebox-cache'
 require 'torquebox-web'
 require 'torquebox-rake-support'
