+++ integration-tests/spec/at_jobs_spec.rb
@@ -1,5 +1,6 @@
 require 'spec_helper'
 require 'torquebox-messaging'
+require 'torquebox-jobs'
 
 remote_describe "at jobs" do
   deploy <<-END.gsub(/^ {4}/,'')
@@ -15,7 +16,7 @@ remote_describe "at jobs" do
     queue = TorqueBox::Messaging::Queue.new('/queue/response')
 
     # Every 440 ms for over 5 seconds, from now
-    TorqueBox::ScheduledJob.at('SimpleJob', :every => 440, :until => Time.now + 5)
+    TorqueBox::Jobs::ScheduledJob.at('SimpleJob', :every => 440, :until => Time.now + 5)
 
     # First run of first spec takes the longest while job runtimes spin up
     count = queue.receive(:timeout => 15_000).nil? ? 0 : 1
@@ -27,14 +28,14 @@ remote_describe "at jobs" do
     # 5 seconds, every 440 ms
     count.should == 12
 
-    TorqueBox::ScheduledJob.remove_sync('SimpleJob').should == true
+    TorqueBox::Jobs::ScheduledJob.remove_sync('SimpleJob').should == true
   end
 
   it "should deploy the :at, :every, :until at job" do
     queue = TorqueBox::Messaging::Queue.new('/queue/response')
 
     # Start in 1 second, then every 460 ms for over 4 seconds (5 seconds from now, but start is delayed)
-    TorqueBox::ScheduledJob.at('SimpleJob', :at => Time.now + 1, :every => 460, :until => Time.now + 5)
+    TorqueBox::Jobs::ScheduledJob.at('SimpleJob', :at => Time.now + 1, :every => 460, :until => Time.now + 5)
 
     # First run takes the longest while job installs
     count = queue.receive(:timeout => 5_000).nil? ? 0 : 1
@@ -46,14 +47,14 @@ remote_describe "at jobs" do
     # 4 seconds, every 460 ms
     count.should == 9
 
-    TorqueBox::ScheduledJob.remove_sync('SimpleJob').should == true
+    TorqueBox::Jobs::ScheduledJob.remove_sync('SimpleJob').should == true
   end
 
   it "should deploy the :in, :repeat, :until at job" do
     queue = TorqueBox::Messaging::Queue.new('/queue/response')
 
     # Start in 1 second, then every 460 ms for over 4 seconds (5 seconds from now, but start is delayed)
-    TorqueBox::ScheduledJob.at('SimpleJob', :in => 1_000, :every => 460, :until => Time.now + 5)
+    TorqueBox::Jobs::ScheduledJob.at('SimpleJob', :in => 1_000, :every => 460, :until => Time.now + 5)
 
     # First run takes the longest while job installs
     count = queue.receive(:timeout => 5_000).nil? ? 0 : 1
@@ -65,14 +66,14 @@ remote_describe "at jobs" do
     # 4 seconds, every 220 ms
     count.should == 9
 
-    TorqueBox::ScheduledJob.remove_sync('SimpleJob').should == true
+    TorqueBox::Jobs::ScheduledJob.remove_sync('SimpleJob').should == true
   end
 
   it "should deploy the :in, :repeat, :every at job" do
     queue = TorqueBox::Messaging::Queue.new('/queue/response')
 
     # Start in 1 second, then repeat te job 10 times, every 150 ms
-    TorqueBox::ScheduledJob.at_sync('SimpleJob', :in => 1_000, :repeat => 10, :every => 150).should == true
+    TorqueBox::Jobs::ScheduledJob.at_sync('SimpleJob', :in => 1_000, :repeat => 10, :every => 150).should == true
 
     # First run takes the longest while job installs
     count = queue.receive(:timeout => 5_000).nil? ? 0 : 1
@@ -84,6 +85,6 @@ remote_describe "at jobs" do
     # 11, because the first execution is not counted
     count.should == 11
 
-    TorqueBox::ScheduledJob.remove_sync('SimpleJob').should == true
+    TorqueBox::Jobs::ScheduledJob.remove_sync('SimpleJob').should == true
   end
 end
