+++ modules/services/src/main/java/org/torquebox/services/RubyServiceMBean.java
@@ -28,6 +28,6 @@ public interface RubyServiceMBean {
     
     String getName();
     String getRubyClassName();
-    String getStatus() throws Exception;
+    String getStatus();
 
 }
